Made WebRequest::getProgress return the number of bytes received.

+++ 3rdparty/libwebrequest/webrequest.h
@@ -7,7 +7,7 @@ struct WebRequest
 {
 	virtual ~WebRequest() { }
 	
-    virtual float getProgress() = 0;
+    virtual int getProgress() = 0;
     virtual bool isDone() = 0;
     virtual bool isSuccess() = 0;
     virtual bool getResultAsData(uint8_t *& bytes, size_t & numBytes) = 0;
